requirements.txt
pdfminer.six==20221105
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3